requirements.txt
streamlit
transformers
torch
torchvision
Pillow
pandas
matplotlib
fastapi
uvicorn
python-multipart